pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
name = "keyfall"
version = "0.1.0"
description = "Open-source piano learning game engine with falling-note visualization"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.11"
dependencies = [
    "pygame>=2.5",
    "mido>=1.3",
    "python-rtmidi>=1.5",
    "pyfluidsynth>=1.3",
    "music21>=9.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.4",
    "pre-commit>=3.0",
]

[project.scripts]
keyfall = "keyfall.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
